Move modals into 'pages/' directory.
