--- resources/assets/components/Modal/index.js
@@ -1,3 +1,0 @@
-export SurveyModalContainer from './containers/SurveyModalContainer';
-export VoterRegistrationModalContainer from './containers/VoterRegistrationModalContainer';
-export PostSignupModal from './containers/PostSignupModalContainer';
