+++ resources/assets/components/Page/CampaignPage/CampaignPage.js
@@ -4,16 +4,13 @@ import PropTypes from 'prop-types';
 import { Switch, Route, Redirect } from 'react-router-dom';
 
 import { FeedContainer } from '../../Feed'; // @TODO: rename to ActivityFeed or ActivityPage...
-import { PostSignupModal } from '../../Modal';
+import PostSignupModalContainer from '../../pages/PostSignupModal/PostSignupModalContainer';
 import BlockPageContainer from '../BlockPage';
 import { isCampaignClosed } from '../../../helpers';
 import { ActionPageContainer } from '../ActionPage';
 import { CampaignSubPageContainer } from '../CampaignSubPage';
 import CampaignFooter from '../../CampaignFooter';
 import Modal from '../../utilities/Modal/Modal';
-
-// TODO: If they click a modal link from the action page, this takes them to the root /.
-// We should probably make a solution that lets them stay on the page they were already at.
 
 const CampaignPage = props => {
   const {
@@ -31,12 +28,9 @@ const CampaignPage = props => {
 
   const isClosed = isCampaignClosed(get(endDate, 'date', null));
 
-  /*
-    If the campaign is closed (and an admin has not specifically
-    toggled the show Action Page button), we want to render the ActivityFeed (community page)
-   *if* it's available (meaning the campaign has an activity feed property populated).
-    Otherwise, we render the ActionPage as usual.
-    */
+  // If the campaign is closed (and an admin has not hit the "Show Action Page" button),
+  // we want to render the community page (activity feed) if it's available, or
+  // if not just render the action page.
   const shouldShowActivityFeed =
     isClosed && !shouldShowActionPage && hasActivityFeed;
   const ActionPageOrActivityFeed = () =>
@@ -51,7 +45,7 @@ const CampaignPage = props => {
       <div>
         {shouldShowSignupAffirmation ? (
           <Modal onClose={clickedHideAffirmation}>
-            <PostSignupModal />
+            <PostSignupModalContainer />
           </Modal>
         ) : null}
         <Switch>
