+++ resources/assets/components/pages/PostSignupModal/PostSignupModal.js
