+++ resources/assets/components/pages/VoterRegistrationModal/VoterRegistrationModal.js
